+++ vite.config.ts
@@ -51,7 +51,11 @@ export default defineConfig({
   test: {
     environment: 'jsdom',
     globals: true,
-    include: ['tests/components/**/*.spec.ts', 'tests/model/**/*.spec.ts'],
+    include: [
+      'tests/components/**/*.spec.ts',
+      'tests/model/**/*.spec.ts',
+      'tests/ssr/**/*.spec.ts',
+    ],
     css: false,
     coverage: {
       provider: 'v8',